# perf(travis) fetch list of jobs a lot faster