Create the job list using the completed job cache only.