@gardleopard 2017-11-07
10x queries using the old code and 10 queries using the new code:

```
for i in $(seq 1 10); do time curl -s localhost:8088/jobs/travis-schibsted ; done 2>&1 >> old.log

real  0m2.890s
user  0m0.000s
sys 0m0.030s

real  0m3.096s
user  0m0.022s
sys 0m0.035s

real  0m2.266s
user  0m0.025s
sys 0m0.000s

real  0m3.097s
user  0m0.000s
sys 0m0.010s

real  0m2.171s
user  0m0.040s
sys 0m0.010s

real  0m2.263s
user  0m0.022s
sys 0m0.000s

real  0m2.553s
user  0m0.031s
sys 0m0.016s

real  0m2.357s
user  0m0.042s
sys 0m0.000s

real  0m3.041s
user  0m0.018s
sys 0m0.017s

real  0m3.340s
user  0m0.000s
sys 0m0.039s


for i in $(seq 1 10); do time curl -s localhost:8088/jobs/travis-schibsted ; done 2>&1 >> new.log

real  0m0.061s
user  0m0.024s
sys 0m0.000s

real  0m0.061s
user  0m0.011s
sys 0m0.011s

real  0m0.080s
user  0m0.021s
sys 0m0.010s

real  0m0.061s
user  0m0.007s
sys 0m0.007s

real  0m0.061s
user  0m0.017s
sys 0m0.000s

real  0m0.065s
user  0m0.021s
sys 0m0.010s

real  0m0.061s
user  0m0.000s
sys 0m0.024s

real  0m0.044s
user  0m0.024s
sys 0m0.000s

real  0m0.081s
user  0m0.018s
sys 0m0.018s

real  0m0.080s
user  0m0.018s
sys 0m0.018s
```