@jervi 2017-11-08
I've tested this as well. Non scientific results below:

  | Old code | New code | Performance increase
-- | -- | -- | --
  | 2s 666ms | 268ms | 895 %
  | 2s 693ms | 334ms | 706 %
  | 2s 711ms | 374ms | 625 %
  | 2s 834ms | 342ms | 729 %
  | 1s 901ms | 348ms | 446 %
  | 1s 35ms | 371ms | 179 %
  | 1s 670ms | 379ms | 341 %
  | 2s 420ms | 531ms | 356 %
  | 3s 933ms | 360ms | 993 %
  | 2s 898ms | 390ms | 643 %
**Average** | **2s 476ms** | **370ms** | **570 %**